Enhance bgfx renderer and fix TGA loading issues
- Implemented color adjustment in the fragment shader to improve visual fidelity.
- Updated SurfaceClass::FindBB to respect input bounds, preventing glyphs from sharing UV regions in the font atlas.
- Fixed TGA loading on Linux by removing a stubbed lowercase compatibility header and ensuring the correct TARGA.H is included.
- Added support for render target readback in bgfx, allowing textures to be read back to the CPU.
- Enhanced texture management by implementing missing DX8Wrapper functions for light state, render targets, and gamma controls.
- Resolved several runtime blockers related to logical audio scene removal and object lifetime management.
- Improved error handling in texture loading and ensured proper resource cleanup.
- Updated documentation to reflect changes and improvements made during the porting process.

+++ Code/wwlib/mpmath.cpp
@@ -2456,7 +2456,7 @@ void MPEXPORT XMP_Randomize(digit * result, Straw & rng, int total_bits, int pre
 	XMP_Init(result, 0, precision);
 	rng.Get(result, nbytes);
 
-	((uint8_t *)result)[nbytes-1] &= (uint8_t)(~((~0) << (total_bits % 8)));
+	((uint8_t *)result)[nbytes-1] &= static_cast<uint8_t>(~((~uint32_t(0)) << (total_bits % 8)));
 }
